% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalgen-package.R
\docType{package}
\name{globalgen-package}
\alias{globalgen}
\alias{globalgen-package}
\title{globalgen: Generates Global Variable Declarations}
\description{
Generate utils::globalVariables() from roxygen tags.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anthonynorth/globalgen}
  \item Report bugs at \url{https://github.com/anthonynorth/globalgen/issues}
}

}
\author{
\strong{Maintainer}: Anthony North \email{anthony.jl.north@gmail.com}

Other contributors:
\itemize{
  \item Miles McBain \email{miles.mcbain@gmail.com} (\href{https://orcid.org/0000-0003-2865-2548}{ORCID}) [contributor]
}

}
\keyword{internal}
