% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_roxyglobals.R
\name{use_roxyglobals}
\alias{use_roxyglobals}
\title{Use roxyglobals}
\usage{
use_roxyglobals()
}
\value{
nothing
}
\description{
Configures roxygen to use \code{\link[=global_roclet]{global_roclet()}}, adds roxyglobals to Suggests
}
\examples{
\dontrun{
use_roxyglobals()
}
}
