% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\alias{options_get_unique}
\alias{options_set_unique}
\alias{options_get_filename}
\alias{options_set_filename}
\title{Roxyglobals options}
\usage{
options_get_unique(file = ".")

options_set_unique(value, file = ".")

options_get_filename(file = ".")

options_set_filename(value, file = ".")
}
\arguments{
\item{file}{DESCRIPTION file to use. By default the DESCRIPTION
file of the current package (i.e. the package the working directory
is part of) is used.}

\item{value}{The new option value}
}
\value{
The option value or nothing
}
\description{
Get and set roxyglobals options in DESCRIPTION file.
}
\section{Functions}{
\itemize{
\item \code{options_get_unique()}: get unique

\item \code{options_set_unique()}: set unique

\item \code{options_get_filename()}: get filename

\item \code{options_set_filename()}: set filename

}}
\examples{
options_get_unique()
\dontrun{
options_set_unique(TRUE)
}
options_get_filename()
\dontrun{
options_set_filename("roxyglobals-generated.R")
}
}
\keyword{internal}
