% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\alias{options_get_unique}
\alias{options_set_unique}
\title{Roxyglobals options}
\usage{
options_get_unique()

options_set_unique(unique)
}
\description{
Get & set roxyglobals options in DESCRIPTION file.
}
\section{Functions}{
\itemize{
\item \code{options_get_unique()}: get unique

\item \code{options_set_unique()}: set unique

}}
\keyword{internal}
